@@ -1,5 +1,5 @@
 {
-    "name": "digisalad-nuxt3-base",
+    "name": "hkrugby-frontend-nuxt",
     "version": "1.0.0",
     "private": true,
     "engines": {
@@ -12,36 +12,48 @@
         "preview": "npx nuxi preview",
         "postinstall": "npx nuxi prepare",
         "analyze": "npx nuxi analyze",
-        "cleanup": "npx nuxi cleanup"
+        "cleanup": "npx nuxi cleanup",
+        "upgrade": "npx nuxi upgrade --force"
     },
     "devDependencies": {
         "@nuxt/devtools": "latest",
         "@nuxtjs/device": "^3.1.0",
+        "@nuxtjs/google-fonts": "^3.1.3",
         "@nuxtjs/i18n": "^8.0.0-rc.4",
-        "@nuxtjs/tailwindcss": "^6.8.0",
-        "@tailwindcss/aspect-ratio": "^0.4.2",
+        "@nuxtjs/sitemap": "^5.0.4",
+        "@unocss/nuxt": "^0.58.0",
+        "@unocss/transformer-directives": "^0.58.0",
         "@vueuse/core": "^10.4.1",
         "@vueuse/nuxt": "^10.4.1",
         "autoprefixer": "^10.4.15",
-        "nuxt": "^3.8.2",
+        "dayjs-nuxt": "^2.1.9",
+        "nuxt": "^3.9.3",
         "nuxt-schema-org": "^3.1.6",
         "nuxt-security": "^0.14.4",
-        "nuxt-simple-robots": "^3.1.2",
-        "nuxt-simple-sitemap": "^3.3.2",
-        "nuxt-site-config": "^1.5.4",
-        "nuxt-site-config-kit": "^1.5.4",
+        "nuxt-simple-robots": "^4.0.0-rc.11",
+        "nuxt-site-config": "^2.2.5",
+        "nuxt-site-config-kit": "^2.2.5",
         "nuxt-viewport": "^2.0.6",
         "sass": "^1.66.1",
         "sass-loader": "^13.3.2",
         "stylelint": "^15.10.3",
         "stylelint-config-standard": "^34.0.0",
-        "vue": "^3.3.4",
-        "vue-router": "^4.2.5"
+        "unocss": "^0.58.0",
+        "vite-plugin-vuetify": "^2.0.1",
+        "vue": "3.4.8",
+        "vue-router": "latest",
+        "vue-tsc": "^1.8.27",
+        "vuetify": "^3.4.9"
     },
     "dependencies": {
+        "@egjs/flicking-plugins": "^4.7.1",
+        "@egjs/vue3-flicking": "^4.11.2",
         "@gtm-support/vue-gtm": "^2.2.0",
-        "@pinia/nuxt": "^0.4.11",
+        "@mdi/font": "^7.2.96",
+        "@pinia/nuxt": "^0.5.1",
         "gsap": "^3.12.2",
-        "pinia": "^2.1.6"
+        "locomotive-scroll": "^5.0.0-beta.11",
+        "pinia": "^2.1.7",
+        "sweetalert2": "^11.7.31"
     }
 }
