@@ -1,4 +1,6 @@
 import { gsap } from 'gsap'
+import colors from '@/vender/unocss/colors.json'
+
 interface AppTransition {
     durationOpen: number //seconds
     durationClose: number //seconds
@@ -6,56 +8,177 @@ interface AppTransition {
 }
 class AppTransition {
     private appTransitionClass = 'ds-app-transition'
-    durationOpen = 0.25
-    durationClose = 0.6
+    private activeClass = 'is-active'
+    durationOpen = 0.8
+    durationClose = 0.7
     constructor(options?: Partial<AppTransition>) {
         Object.assign(this, options)
 
-        //process.client && this.init()
+        process.client && this.init()
     }
 
     //取得動畫容器
     private get container(): HTMLElement | null {
-        const c = document.querySelector(`.${this.appTransitionClass}`) as any
+        const _this = this
+        const c = document.querySelector(`.${_this.appTransitionClass}`) as any
         return c
     }
 
-    public init = () => {
-        const el = this.container
-        el?.style.setProperty('--duration_open', `${this.durationOpen}s`)
-        el?.style.setProperty('--duration_close', `${this.durationClose}s`)
+    private init = () => {
+        const _this = this
+        // const el = _this.container
+        // el?.style.setProperty('--duration_open', `${_this.durationOpen}s`)
+        // el?.style.setProperty('--duration_close', `${_this.durationClose}s`)
     }
 
     //過場動畫:開始
     public open = () => {
-        const el = this.container
+        const _this = this
+        const el = _this.container
+        const board = el?.querySelector(`.t-borad`) as HTMLElement
+        const locoInner = document.querySelector('.locomotive-layout') as HTMLElement
+        const header = document.querySelector('header')
+        const clipPath = () => {
+            const winW = window.innerWidth
+
+            return `polygon(0 100%, ${winW * 0.2}px 0%, 100% 0%, 100% 100%)`
+        }
+
+        const resetCircleOptions = {
+            alpha: 0,
+            scaleX: 0.25,
+            transformOrigin: '100% 50%',
+            duration: 0,
+            backgroundColor: colors.blue.primary,
+        }
+        const resetLocoInnerOptions = {
+            alpha: 1,
+            duration: 0,
+        }
         const _t = gsap.timeline({
             paused: true,
-            duration: this.durationOpen,
-            onStart: () => {
-                el?.classList.add('is-active')
+            onStart: function () {
+                el?.classList.add(_this.activeClass)
+            },
+            onReverseComplete: function () {
+                el?.classList.remove(_this.activeClass)
             },
         })
-        //.set(el, { visibility: 'visible', autoAlpha: 0 })
-        //.to(el, { autoAlpha: 1 })
+        _t.set(board, resetCircleOptions)
+        _t.set(locoInner, resetLocoInnerOptions)
+        if (!locoInner) {
+            _t.totalDuration(0)
+            return _t
+        }
+        _t.set(board, resetCircleOptions)
+            .set([locoInner], resetLocoInnerOptions)
+            .to(board, { alpha: 1, ease: 'none', duration: _this.durationOpen * 0.2 }, 0)
+            .to(
+                board,
+                {
+                    scaleX: 1,
+                    backgroundColor: colors.blue.primary,
+                    ease: 'power2.inOut',
+                },
+                0,
+            )
+            .to(
+                board,
+                {
+                    keyframes: {
+                        '0%': {
+                            clipPath,
+                        },
+                        '100%': {
+                            clipPath: 'polygon(0 100%, 0% 0%, 100% 0%, 100% 100%)',
+                        },
+                    },
+                    ease: 'power2.inOut',
+                    duration: _this.durationOpen * 0.9,
+                },
+                0,
+            )
+            .to(
+                [locoInner, header],
+                {
+                    alpha: 0,
+                    ease: 'power2.inOut',
+                },
+                0,
+            )
 
+        _t.totalDuration(_this.durationOpen)
         return _t
     }
 
     //過場動畫:結束
     public close = () => {
-        const el = this.container
+        const _this = this
+        const el = _this.container
+        const board = el?.querySelector(`.t-borad`) as HTMLElement
+        const locoInner = document.querySelector('.locomotive-layout') as HTMLElement
+        const header = document.querySelector('header')
+
+        const clipPathReset = () => {
+            return `polygon(0 100%, 0% 0%, 100% 0%, 100% 100%)`
+        }
+        const clipPath = () => {
+            const winW = window.innerWidth
+            const p = 0.5
+            const point = ((winW * (1 - p)) / winW) * 100
+            return `polygon(0 100%, 0% 0%, 100% 0%, ${point}% 100%)`
+        }
+
+        const resetCircleOptions = {
+            transformOrigin: '0 50%',
+            xPercent: 0,
+            scaleX: 1,
+            alpha: 1,
+            backgroundColor: colors.blue.primary,
+        }
+
+        //let isTriggerFinished = false
         const _t = gsap.timeline({
             paused: true,
-            duration: this.durationClose,
-            onStart: () => {
-                el?.classList.remove('is-active')
+            onStart: function () {
+                el?.classList.add(_this.activeClass)
             },
-            // onComplete: () => {
-            //     gsap.set(el, { visibility: 'hidden' })
-            // },
+            onComplete: function () {
+                el?.classList.remove(_this.activeClass)
+                gsap.set(locoInner, { clearProps: 'all' })
+            },
         })
-        //.to(el, { autoAlpha: 0})
+        if (!locoInner) {
+            _t.totalDuration(0)
+            return _t
+        }
+        _t.set(board, resetCircleOptions)
+            .set([locoInner], {
+                alpha: 0,
+            })
+            .to(
+                board,
+                {
+                    keyframes: {
+                        '0%': {
+                            clipPath: clipPathReset,
+                        },
+                        '90%': {
+                            clipPath,
+                        },
+                        '100%': {
+                            clipPath: clipPathReset,
+                            ease: 'power2.in',
+                        },
+                    },
+                },
+                0,
+            )
+            .to(board, { scaleX: 0, ease: 'power2.inOut' }, 0)
+            .to([locoInner, header], { alpha: 1, ease: 'power2.inOut' }, 0)
+
+        _t.totalDuration(_this.durationClose)
+
         return _t
     }
 }
