@@ -1,5 +1,8 @@
+import { useGlobalStore } from '~/store'
+
 export default defineNuxtPlugin((nuxtApp) => {
-    const { toggleScreenResizeComplete } = useAfterScreenResize()
+    const storeGlobal = useGlobalStore()
+    const { toggleScreenResizeComplete, toggleScreenResizing } = useAfterScreenResize()
     const { width, height } = useWindowSize()
     const windowObj = reactive({ width, height })
 
@@ -8,6 +11,7 @@ export default defineNuxtPlugin((nuxtApp) => {
     const resizeListen = () => {
         const { width, height } = useWindowSize()
         clearTimeout(resizeTimer)
+        toggleScreenResizing()
         if (width !== windowWidth) {
             toggleScreenResizeComplete(false)
             windowWidth = useWindowSize().width
@@ -18,4 +22,14 @@ export default defineNuxtPlugin((nuxtApp) => {
     }
     window.addEventListener('resize', resizeListen)
     //#endregion
+
+    // 監聽是否正在瀏覽網頁
+    window.onpagehide = window.onblur = () => {
+        storeGlobal.isPageFocus = false
+    }
+    window.onpageshow = window.onfocus = () => {
+        storeGlobal.isPageFocus = true
+    }
+
+    window.globalTimeout = []
 })
